refactor: landing page responsive screen
