+++ apps/web/app/(home)/layout.tsx
@@ -1,6 +1,6 @@
 export default function HomeLayout({ children }: { children: React.ReactNode }) {
   return (
-    <div className="h-screen w-full pt-[142px] md:pl-[45px] md:pr-0 xl:flex xl:h-[calc(100vh-78px)] xl:flex-col xl:justify-center xl:px-[30px] 2xl:px-[150px]">
+    <div className="flex h-[calc(100vh-78px)] min-h-[720px] w-full justify-center pt-[78px] md:min-h-[920px] md:justify-between md:pl-[45px] md:pr-0 xl:flex-col xl:items-center xl:justify-center xl:px-[30px]">
       {children}
     </div>
   );
